moved the bd to its own dir
