+++ src/models/userModel.js
@@ -1,5 +1,5 @@
 // Example user model (assuming you'll use a database later)
-const db = require('../database');
+const db = require('../database/database');
 
 class User {
   constructor(id, name, email) {
